+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSessionManagerImpl.java
@@ -199,15 +199,14 @@ public abstract class ReplicantSessionManagerImpl
    * @param session   the session.
    * @param etag      the etag for message if any.
    * @param changeSet the messages to be sent along to the client.
-   * @return the packet created.
    */
-  protected Packet sendPacket( @Nonnull final ReplicantSession session,
-                               @Nullable final String etag,
-                               @Nonnull final ChangeSet changeSet )
+  protected void sendPacket( @Nonnull final ReplicantSession session,
+                             @Nullable final String etag,
+                             @Nonnull final ChangeSet changeSet )
   {
     final Integer requestId = (Integer) getRegistry().getResource( ServerConstants.REQUEST_ID_KEY );
     getRegistry().putResource( ServerConstants.REQUEST_COMPLETE_KEY, Boolean.FALSE );
-    return session.getQueue().addPacket( requestId, etag, changeSet );
+    session.sendPacket( requestId, etag, changeSet );
   }
 
   /**